Refactors top-level robosat package into sub-packages, closes #2
