+++ robosat/tools/subset.py
@@ -4,7 +4,7 @@ import shutil
 
 from tqdm import tqdm
 
-from robosat.tiles import tiles_from_slippy_map, tiles_from_csv
+from robosat.geo.tiles import tiles_from_slippy_map, tiles_from_csv
 
 
 def add_parser(subparser):
